+++ src/features/api/index.ts
@@ -1,10 +1,10 @@
 import cors from "cors";
-import { type BrowserWindow, dialog } from "electron";
+import { dialog, type BrowserWindow } from "electron";
 import express from "express";
 import swaggerUi from "swagger-ui-express";
 
 import { settings } from "../../constants/settings";
-import { settingsStore } from "../../scripts/settings";
+import { settingsStore } from "../../scripts/settingsStore";
 import { addCurrentInfo } from "./features/current";
 import { addHealthEndpoint } from "./features/health";
 import { addPlaybackControl } from "./features/player";